requirements.txt
crewai>=0.21.0
crewai_tools
requests
python-dotenv
pandas
torch
nltk
openpyxl
pyyaml